(+) support debug of `conan build ...`
In VSCode, you can now debug the run of Conan to build the project.
Choose `conanfile.py` in the combo box, at the top of the Debugger pane.
You can put breakpoints in `conanfile.py` to track Conan's build
process.
Change the command line params in `launch.json`.
The build process is done in `build.sh` but it is really just Conan's
[Package Development Flow](https://docs.conan.io/2/tutorial/developing_packages/local_package_development_flow.htm).