(-) fix creation of compile_commands.json
1. fix the creation of compile_commands.json
for some reason, setting the option in the root's CMakeLists.txt
failed on a fresh Ubuntu 22.04 docker image.
the fix is to set it in the base.profile
2. since clangd doesn't find compile_commands.json where it is created,
I had to copy it to the build folder.
I change it here to be a symlink which update when we change the build
dir - debug/release, profile changes.

I think we deserve better, but this is what I got now :-|